+++ financial-control/src/main/java/larissagalao/com/github/financialcontrol/repositories/BillsRepository.java
@@ -16,7 +16,7 @@ public interface BillsRepository extends JpaRepository<Bills, Integer> {
     Integer total();
 
     @Query(value = "SELECT sum(valor) FROM Bills where tipo=?1")
-    Integer totalbyType(String string);
+    Double totalbyType(String string);
 
     @Query(value = "SELECT tipo FROM Bills")
     List<String> types();
